backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase>=2.3.0
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
prophet==1.1.5
scikit-learn>=1.4.0
pydantic>=2.5.0
python-multipart==0.0.6
openpyxl>=3.1.2
httpx>=0.26.0
numbers-parser>=4.0.0
